+++ src/resources/config/module.php
@@ -6,6 +6,7 @@ return [
     'auto_destroy'      => false, // Whether to immediately delete carts with 0 items
     'auto_assign_user'  => true, // Whether to automatically set the user_id on new carts (based on Auth::user())
     'preserve_for_user' => false, // Whether to keep and restore user carts across logins and devices
+    'merge_duplicates'  => false, // Whether to merge carts if `preserve_for_user` is enabled, user logs in and the session contains another cart
     'items'             => [
         'extra_product_attributes_to_merge' => []
     ]